move epic look up to id lookup
