+++ src/dodal/beamlines/i06_shared.py
@@ -39,7 +39,7 @@ def pgm() -> PlaneGratingMonochromator:
 # Insertion Device
 # ------------- Downstream Insertion Device --------------------
 @devices.factory()
-def i06_idd_epics_polynomial_device() -> I06EpicsPolynomialDevice:
+def idd_epics_polynomial_device() -> I06EpicsPolynomialDevice:
     return I06EpicsPolynomialDevice(prefix=f"{PREFIX.beamline_prefix}-OP-IDD-01:")
 
 
@@ -65,20 +65,20 @@ def idd(idd_gap: UndulatorGap, idd_phase: UndulatorLockedPhaseAxes) -> Apple2:
 
 @devices.factory()
 def idd_controller(
-    idd: Apple2, i06_idd_epics_polynomial_device: I06EpicsPolynomialDevice
+    idd: Apple2, idd_epics_polynomial_device: I06EpicsPolynomialDevice
 ) -> I06Apple2Controller:
     """I06 downstream insertion device controller."""
     return I06Apple2Controller(
         apple2=idd,
-        gap_energy_motor_lut=i06_idd_epics_polynomial_device.energy_gap_motor_lookup,
-        phase_energy_motor_lut=i06_idd_epics_polynomial_device.energy_phase_motor_lookup,
-        inverse_gap_energy_motor_lut=i06_idd_epics_polynomial_device.gap_motor_energy_lookup,
+        gap_energy_motor_lut=idd_epics_polynomial_device.energy_gap_motor_lookup,
+        phase_energy_motor_lut=idd_epics_polynomial_device.energy_phase_motor_lookup,
+        inverse_gap_energy_motor_lut=idd_epics_polynomial_device.gap_motor_energy_lookup,
     )
 
 
 # -------------------- Upstream Insertion Device -------------------
 @devices.factory()
-def i06_idu_epics_polynomial_device() -> I06EpicsPolynomialDevice:
+def idu_epics_polynomial_device() -> I06EpicsPolynomialDevice:
     return I06EpicsPolynomialDevice(prefix=f"{PREFIX.beamline_prefix}-OP-IDU-01:")
 
 
@@ -104,14 +104,14 @@ def idu(idu_gap: UndulatorGap, idu_phase: UndulatorLockedPhaseAxes) -> Apple2:
 
 @devices.factory()
 def idu_controller(
-    idu: Apple2, i06_idu_epics_polynomial_device: I06EpicsPolynomialDevice
+    idu: Apple2, idu_epics_polynomial_device: I06EpicsPolynomialDevice
 ) -> I06Apple2Controller:
     """I06 upstream insertion device controller."""
     return I06Apple2Controller(
         apple2=idu,
-        gap_energy_motor_lut=i06_idu_epics_polynomial_device.energy_gap_motor_lookup,
-        phase_energy_motor_lut=i06_idu_epics_polynomial_device.energy_phase_motor_lookup,
-        inverse_gap_energy_motor_lut=i06_idu_epics_polynomial_device.gap_motor_energy_lookup,
+        gap_energy_motor_lut=idu_epics_polynomial_device.energy_gap_motor_lookup,
+        phase_energy_motor_lut=idu_epics_polynomial_device.energy_phase_motor_lookup,
+        inverse_gap_energy_motor_lut=idu_epics_polynomial_device.gap_motor_energy_lookup,
     )
