+++ src/dodal/devices/beamlines/i06_shared/__init__.py
@@ -1,10 +1,9 @@
 from .i06_apple2_controller import I06Apple2Controller
 from .i06_enum import I06Grating
-from .i06_epics_lookup import EpicsPolynomialEnergyMotorLookup, I06EpicsPolynomialDevice
+from .i06_epics_lookup import I06EpicsPolynomialDevice
 
 __all__ = [
     "I06Grating",
     "I06EpicsPolynomialDevice",
     "I06Apple2Controller",
-    "EpicsPolynomialEnergyMotorLookup",
 ]
