+++ src/dodal/devices/insertion_device/apple2_controller.py
@@ -18,9 +18,7 @@ from dodal.devices.insertion_device.apple2_undulator import (
     PhaseAxesType,
     UndulatorPhaseAxes,
 )
-from dodal.devices.insertion_device.energy_motor_lookup import (
-    EnergyMotorLookup,
-)
+from dodal.devices.insertion_device.energy_motor_lookup import EnergyMotorLookup
 from dodal.devices.insertion_device.enum import Pol
 from dodal.log import LOGGER
